refactor(models): promote primitives.py to a per-concept package (#161)
Splits the 273-line models/primitives.py into one-concept-per-module
(fqdn, http_url, ip_addr, cidr, email, ip_type) under models/primitives/,
mirroring the models/nvd/ and models/mitre/ layout. The package __init__
re-exports every primitive, so 'from models.primitives import FQDN' and the
top-level 'from models import FQDN' keep working unchanged - the move is
import-transparent to all ~20 consumers. The one intra-package edge is
http_url -> fqdn (a URL's host runs through the FQDN validator).

Docstrings and comments are carried over verbatim. The non-str guards in
_validate_fqdn / _validate_endpoint_url gain the same
'# pragma: no cover - Pydantic enforces str upstream' the three sibling
validators (ip_addr/email/cidr) already carry: the branch is unreachable
through the Annotated[str, ...] boundary. They escaped notice before only
because diff-cover never examined pre-existing lines; the move surfaced the
inconsistency. Updates the two stale path references (endpoint.py docstring,
test_primitives.py docstring). Full clean stack verified green against a
freshly fetched origin/main.
